requirements.txt
streamlit==1.48.0
openai==1.40.0
gTTS==2.5.4
Pillow==11.3.0